% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlin_wfs.R
\name{berlin_wfs}
\alias{berlin_wfs}
\title{Compute descriptive statistics for the numeric columns of a data frame.}
\usage{
berlin_wfs(table, Geo_path, ...)
}
\arguments{
\item{table}{The data frame where you find the names of the file, the data and the url.}

\item{Geo_path}{The path where the data has to be stored}

\item{...}{Optional. Columns in the data frame}
}
\value{
A folder with the data as a geopackage.
}
\description{
Compute descriptive statistics for the numeric columns of a data frame.
}
\examples{
\dontrun{
berlin_wfs(table = table, Geo_path = ".")
}
}
