% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_fisbroker.R
\name{download_fisbroker}
\alias{download_fisbroker}
\title{Download data from fisbroker}
\usage{
download_fisbroker(link)
}
\arguments{
\item{link}{the url of the data from fisbroker.}
}
\value{
A folder with the data as a geopackage.
}
\description{
Download data from fisbroker
}
\examples{
\dontrun{
download_fisbroker()
}
}
