% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_fis_tab.R
\name{show_fis_tab}
\alias{show_fis_tab}
\title{show the names and links of the table for downloading the wfs layers.}
\usage{
show_fis_tab(x = "./data-raw/fis_broker_wfs_links.csv")
}
\arguments{
\item{x}{The path where the data lays}
}
\value{
A dataframe with all the names and links.
}
\description{
show the names and links of the table for downloading the wfs layers.
}
\examples{
\dontrun{
show_fis_tab()
}
}
